move operatepod package to utilsk8s
