+++ internal/utilk8s/operatepod.go
@@ -1,4 +1,4 @@
-package operatepod
+package utilk8s
 
 import (
 	"context"
